Remove Jekyll, leave generated static HTML
Time flows past 2019; and Jekyll project,
its plugins and various other dependencies used
to build this site have become brittle. The build will not work
for indefinate amount of time.

To preserve the site for histocial content, remove all of the
dynamic build and generation, and leave only static HTML.
Site is not expected to change further.

If Tilt 20XX happens, the history is still in Git for restoration.
TILT